requirements.txt
streamlit~=1.28.2
cohere~=4.34
python-dotenv~=1.0.0